+++ src/index.js
@@ -28,8 +28,8 @@ router();
 window.addEventListener('load', router);
 window.addEventListener('hashchange', router);
 
-alert(
-  'Нажаль ми не встигли доробити до кінця 💔   Дуже просимо, по можливості, перевірити у четверг 12.01.2023 😘 Дякуємо за розуміння!'
-);
+// alert(
+//   'Нажаль ми не встигли доробити до кінця 💔   Дуже просимо, по можливості, перевірити у четверг 12.01.2023 😘 Дякуємо за розуміння!'
+// );
 
-console.log("Link to Pull Request",'https://github.com/Katerina-Zamiatina/online-shop/pull/21');
+// console.log("Link to Pull Request",'https://github.com/Katerina-Zamiatina/online-shop/pull/21');
